refactor(caret): integrate user selection tracking directly into CaretModel
- Removed UserSelectingTracker class and its associated tests, consolidating user selection logic within CaretModel.
- Implemented a new #trackUserSelecting method to handle mouse events for user selection tracking.
- Updated isUserSelecting property to initialize directly in CaretModel, simplifying the selection state management.
