+++ packages/core/src/features/caret/index.ts
@@ -1,4 +1,3 @@
 export {CaretModel} from './CaretModel'
 export {TriggerFinder} from './TriggerFinder'
 export * as caretDom from './caretDom'
-export {UserSelectingTracker} from './UserSelectingTracker'
